Report and respond to CustomerWindow.isDebugged
If the "Login" Debugging Check detected unauthorized use of a
debugger, report the incident to Application Insights when the
CustomerWindow is constructed.

Then, in two event handlers, throw a meaningless exception to stop
a session compromised by unauthorized debugging.

By separating the detection, reporting, and reaction, we make it
harder for an attacker to work around the Check.