Configure "Query" Debugging Check
This Check will detect unauthorized use of a debugger when any of the
following methods are called:

* CreditCardWindow.UpdateData
* EmailWindow.UpdateData
* PhoneWindow.UpdatePhones

It will call the ReportDebugging method on the corresponding type and
instance of the method that was called, providing as the argument
`true` if a debugger was detected and `false` otherwise.

The Check will then immediately exit the application if a debugger was
detected.