Configure the Tamper Check
This Check will detect if the application binary has been modified
when any of the following methods are called:

* CustomerWindow.UpdateData
* LoginDialog.ConfirmLogin
* Utilities.ShowAndHandleDialog

It will call the Utilities.ReportTampering method, providing as the
argument `true` if tampering was detected and `false` otherwise.

The Check will then immediately exit the application if tampering was
detected.